taylor|style(minor-adjustments):  - Refactor code move types and util methods to separate files
